Rename `acting_user` to `user` in policies
Definite tradeoffs in both directions but after a group discussion, felt
that new engineers having to investigate/learn another term wasn't worth
the benefits of the explicit qualification. Dropping back to the term
used in pundit docs/examples with the understanding that any _other_
users in play can be qualified as needed.